Document + refactor QuestNameToHelper
